moved UpdateServerAddress and UpdateTimeout
