+++ client/client.go
@@ -9,7 +9,6 @@ import (
 	"net/http"
 	"os"
 	"strings"
-	"time"
 
 	ofcmessages "github.com/thomas-osgood/ofs/client/internal/messages"
 	ofscommon "github.com/thomas-osgood/ofs/internal/general"
@@ -241,23 +240,6 @@ func (fc *FClient) RenameFile(originalname string, newname string) (err error) {
 		return fmt.Errorf(status.GetMessage())
 	}
 
-	return nil
-}
-
-// function designed to update the address of the server
-// the client attempts to connect to when it makes RPC calls.
-func (fc *FClient) UpdateServerAddress(newaddress string) (err error) {
-	fc.srvaddr = newaddress
-	return nil
-}
-
-// function designed to update the connection timeout for
-// the client.
-func (fc *FClient) UpdateTimeout(newtimeout time.Duration) (err error) {
-	if newtimeout.Seconds() <= 0 {
-		return fmt.Errorf(ofcmessages.ERR_NEGATIVE_TIMEOUT)
-	}
-	fc.timeout = newtimeout
 	return nil
 }
